Logica de precos + testes/ajustes

+++ estapar-garage-webhook/src/main/java/br/com/estapar/garage/webhook/model/dto/OccupanceBalanceDTO.java
@@ -8,5 +8,5 @@ import lombok.Data;
 public class OccupanceBalanceDTO {
 
     private String sector;
-    private Long totalOccupance;
+    private double totalOccupance;
 }
